requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
biopython==1.83
biotite==0.39.0
bitsandbytes==0.43.0
bleach==6.1.0
blessed==1.20.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.17.1
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.14.0
defusedxml==0.7.1
dill==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
exceptiongroup==1.2.0
executing==2.0.1
fair-esm==2.0.0
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
ginnm==0.0.2
gitdb==4.0.11
GitPython==3.1.42
gmpy2==2.1.2
gpustat==1.1.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
idna==3.4
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
javalang==0.13.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lightning==2.2.0.post0
lightning-utilities==0.10.1
llvmlite==0.42.0
lmdb==1.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
notebook==7.1.1
notebook_shim==0.2.4
numba==0.59.0
numpy==1.26.4
nvidia-ml-py==12.535.133
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
pathos==0.3.2
peft==0.9.0
pexpect==4.9.0
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
pox==0.3.4
ppft==1.7.6.8
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-rouge==1.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyg-lib==0.4.0+pt21cu118
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.9.0
python-json-logger==2.0.7
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
sentry-sdk==1.40.6
setproctitle==1.3.3
setuptools==68.2.2
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
terminado==0.18.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.1.1
torch-cluster==1.6.3+pt21cu118
torch_geometric==2.5.0
torch-scatter==2.1.2+pt21cu118
torch-sparse==0.6.18+pt21cu118
torch-spline-conv==1.2.2+pt21cu118
torchaudio==2.1.1
torchmetrics==1.3.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
triton==2.1.0
trl==0.8.1
types-python-dateutil==2.8.19.20240311
typing_extensions==4.9.0
tyro==0.7.3
uri-template==1.3.0
urllib3==2.1.0
wandb==0.16.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.41.2
widgetsnbextension==4.0.10
xxhash==3.4.1
yarl==1.9.4
